+++ src/Writer.php
@@ -14,6 +14,7 @@ namespace League\Csv;
 
 use InvalidArgumentException;
 use League\Csv\Modifier\RowFilter;
+use League\Csv\Modifier\StreamIterator;
 use ReflectionMethod;
 use SplFileObject;
 use Traversable;
@@ -42,38 +43,18 @@ class Writer extends AbstractCsv
     protected $csv;
 
     /**
-     * fputcsv method from SplFileObject
+     * fputcsv method from SplFileObject or StreamIterator
      *
      * @var ReflectionMethod
      */
-    protected static $fputcsv;
+    protected $fputcsv;
 
     /**
      * Nb parameters for SplFileObject::fputcsv method
      *
      * @var integer
      */
-    protected static $fputcsv_param_count;
-
-    /**
-     * @inheritdoc
-     */
-    protected function __construct($path, $open_mode = 'r+')
-    {
-        parent::__construct($path, $open_mode);
-        static::initFputcsv();
-    }
-
-    /**
-     * initiate a SplFileObject::fputcsv method
-     */
-    protected static function initFputcsv()
-    {
-        if (null === static::$fputcsv) {
-            static::$fputcsv = new ReflectionMethod('\SplFileObject', 'fputcsv');
-            static::$fputcsv_param_count = static::$fputcsv->getNumberOfParameters();
-        }
-    }
+    protected $fputcsv_param_count;
 
     /**
      * Adds multiple lines to the CSV document
@@ -115,18 +96,38 @@ class Writer extends AbstractCsv
         }
         $row = $this->formatRow($row);
         $this->validateRow($row);
-
-        if (null === $this->csv) {
-            $this->csv = $this->getIterator();
-        }
-
-        static::$fputcsv->invokeArgs($this->csv, $this->getFputcsvParameters($row));
-        if ("\n" !== $this->newline) {
-            $this->csv->fseek(-1, SEEK_CUR);
-            $this->csv->fwrite($this->newline);
-        }
+        $this->addRow($row);
 
         return $this;
+    }
+
+    /**
+     * Add new record to the CSV document
+     *
+     * @param array $row record to add
+     */
+    protected function addRow(array $row)
+    {
+        $this->initCsv();
+        $this->fputcsv->invokeArgs($this->csv, $this->getFputcsvParameters($row));
+        if ("\n" !== $this->newline) {
+            $this->csv->fseek(-1, SEEK_CUR);
+            $this->csv->fwrite($this->newline, strlen($this->newline));
+        }
+    }
+
+    /**
+     * Initialize the CSV object and settings
+     */
+    protected function initCsv()
+    {
+        if (null !== $this->csv) {
+            return;
+        }
+
+        $this->csv = $this->getIterator();
+        $this->fputcsv = new ReflectionMethod(get_class($this->csv), 'fputcsv');
+        $this->fputcsv_param_count = $this->fputcsv->getNumberOfParameters();
     }
 
     /**
@@ -139,7 +140,7 @@ class Writer extends AbstractCsv
     protected function getFputcsvParameters(array $fields)
     {
         $parameters = [$fields, $this->delimiter, $this->enclosure];
-        if (4 == static::$fputcsv_param_count) {
+        if (4 == $this->fputcsv_param_count) {
             $parameters[] = $this->escape;
         }