+++ test/Plugin/NullValidatorTest.php
@@ -4,15 +4,15 @@ namespace League\Csv\Test\Plugin;
 
 use League\Csv\Exception\InvalidRowException;
 use League\Csv\Plugin\ForbiddenNullValuesValidator;
-use League\Csv\Test\AbstractTestCase;
 use League\Csv\Writer;
+use PHPUnit_Framework_TestCase;
 use SplFileObject;
 use SplTempFileObject;
 
 /**
  * @group validators
  */
-class NullValidatorTest extends AbstractTestCase
+class NullValidatorTest extends PHPUnit_Framework_TestCase
 {
     private $csv;